+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # v1.18.0 - ???
 
 ### Features
+- Added the !sc party command with more detailed info about scs
+- Added the !drop party command with detailed info about rare drops and dyes
 
 ### Fixes
 - Added Elementa memory leak patch (toggleable in Other settings if it causes conflicts)
 
 ### Changes
-- Changed the !since command to display sea creatures caught since last rare drop and catches since last sea creature
+- Changed the !since command to display only simple info
 
 ### Back-end